+++ bundles/digiroad/digiroad-myplaces2/request/EditRequestHandler.js
@@ -26,12 +26,11 @@ Oskari.clazz.define('Oskari.digiroad.bundle.myplaces2.request.EditRequestHandler
          *      request to handle
          */
         handleRequest: function (core, request) {
-            var sandbox = core.getSandbox();
             if (request.getName() == 'DigiroadMyPlaces.EditPlaceRequest') {
-                this._handleEditPlace(sandbox, request);
+                this._handleEditPlace(request);
             }
         },
-        _handleEditPlace: function (sandbox, request) {
+        _handleEditPlace: function (request) {
             var service = this.instance.getService(),
                 place = service.findMyPlace(request.getId());
             if (place) {