+++ bundles/mapping/mapmodule/plugin/publishertoolbar/request/ToolContainerRequestHandler.js
@@ -23,7 +23,6 @@ function(toolbarPlugin) {
      *      request to handle
      */
     handleRequest : function(core, request) {
-        var sandbox = core.getSandbox();
         if(request.getOp() == 'set') {
             this._toolbar.setToolContent(request.getData());
         }