+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 - Oskari.mapframework.sandbox.Sandbox has been renamed Oskari.Sandbox
 - Moved domain/Map from core to sandbox.
 - Moved getRequestParameter() from core and sandbox to Oskari.util.getRequestParam()
+- Removed core.getSandbox(). Use Oskari.getSandbox() instead.
 
 #### Request/Event refactoring
 - Moved files from under sources to mapmodule: MapMoveRequest, AfterMapMoveEvent, MapMoveStartEvent and MouseHoverEvent